requirements.txt
psycopg2
load_dotenv
os
pandas
streamlit
supabase